Fixes for policy-metrics and schema subcommands based on code review (#593)
