+++ c7n/schema.py
@@ -37,7 +37,7 @@ from jsonschema.exceptions import best_match
 from c7n.manager import resources
 from c7n.resources import load_resources
 from c7n.filters import ValueFilter, EventFilter, AgeFilter
-from c7n.utils import ArgumentError
+from c7n.exceptions import ArgumentError
 
 
 def validate(data, schema=None):
@@ -372,97 +372,3 @@ def json_dump(resource=None):
         traceback.print_exc()
         pdb.post_mortem(sys.exc_info()[-1])
 
-
-def print_schema(options):
-    """
-    Print information about the schema.
-    """
-    if options.json:
-        json_dump(options.resource)
-        return
-        
-    load_resources()
-    resource_mapping = resource_vocabulary()
-
-    if options.summary:
-        schema_summary(resource_mapping)
-        return
-
-    # Here are the formats for what we accept:
-    # - No argument
-    #   - List all available RESOURCES
-    # - RESOURCE
-    #   - List all available actions and filters for supplied RESOURCE
-    # - RESOURCE.actions
-    #   - List all available actions for supplied RESOURCE
-    # - RESOURCE.actions.ACTION
-    #   - Show class doc string and schema for supplied action
-    # - RESOURCE.filters
-    #   - List all available filters for supplied RESOURCE
-    # - RESOURCE.filters.FILTER
-    #   - Show class doc string and schema for supplied filter
-
-    if not options.resource:
-        resource_list = {'resources': sorted(resources.keys()) }
-        print(yaml.safe_dump(resource_list, default_flow_style=False))
-        return
-
-    # Format is RESOURCE.CATEGORY.ITEM
-    components = options.resource.split('.')
-
-    #
-    # Handle resource
-    #
-    resource = components[0].lower()
-    if resource not in resource_mapping:
-        raise ArgumentError('{} is not a valid resource'.format(resource))
-
-    if len(components) == 1:
-        del(resource_mapping[resource]['docs'])
-        output = {resource: resource_mapping[resource]}
-        print(yaml.safe_dump(output))
-        return
-
-    #
-    # Handle category
-    #
-    category = components[1].lower()
-    if category not in ('actions', 'filters'):
-        raise ArgumentError("Valid choices are 'actions' and 'filters'.  You supplied '{}'".format(category))
-    
-    if len(components) == 2:
-        output = "No {} available for resource {}.".format(category, resource)
-        if category in resource_mapping[resource]:
-            output = {resource: {category: resource_mapping[resource][category]}}
-        print(yaml.safe_dump(output))
-        return
-
-    #
-    # Handle item
-    #
-    item = components[2].lower()
-    if item not in resource_mapping[resource][category]:
-        raise ArgumentError('{} is not in the {} list for resource {}'.format(item, category, resource))
-
-    if len(components) == 3:
-        docstring = resource_mapping[resource]['docs'][category][item]
-        if docstring:
-            print(docstring)
-        else:
-            print("No help is available for this item.")
-        return
-
-    # We received too much (e.g. s3.actions.foo.bar)
-    raise ArgumentError("Invalid selector '{}'.  Max of 3 components in the "\
-                        "format RESOURCE.CATEGORY.ITEM".format(options.resource))
-
-
-if __name__ == '__main__':
-    # dump our schema
-    #
-    # The canonical way to do this now is `custodian schema --json`, but I am
-    # leaving in this __main__ section in case people are relying on the old
-    # ability to just do:
-    # 
-    # $ python -m c7n.schema
-    json_dump()
